Move MainLayout to components
