+++ frontend/pages/como-funciona.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Layout from "./MainLayout";
+import Layout from "../components/MainLayout";
 
 export default function ComoFunciona() {
   return (
@@ -11,5 +11,4 @@ export default function ComoFunciona() {
         {/* TODO: Adicione aqui os passos do processo */}
       </main>
     </Layout>
-  );
-}
+  );}
